requirements.txt
paddlepaddle==2.6.2
paddleocr==2.9.1
pypinyin
pykakasi
korean-romanizer
deep-translator
opencv-python<4.11
numpy<2
Pillow
PyQt5
pytest
manga-ocr
transformers<4.45
